Extract the "SearchArray block out of Search and into it's own class
